ignore /Directory.Build.targets use shared project for shared functionality update Gt7,TDUSC, Overload and ACE plugin to use Shared library
