+++ ACEPlugin/ACEPlugin.csproj
@@ -29,4 +29,5 @@
   <ItemGroup>
     <ProjectReference Include="..\YawGLAPI\YawGLAPI.csproj" />
   </ItemGroup>
+  <Import Project="..\SharedLib\SharedLib.projitems" Label="Shared" />
 </Project>
