+++ GT7Plugin/GT7Plugin.csproj
@@ -32,4 +32,6 @@
     <ProjectReference Include="..\YawGLAPI\YawGLAPI.csproj" />
   </ItemGroup>
 
+  <Import Project="..\SharedLib\SharedLib.projitems" Label="Shared" />
+
 </Project>
