+++ OverloadPlugin/OverloadPlugin.csproj
@@ -31,4 +31,5 @@
   <ItemGroup>
     <ProjectReference Include="..\YawGLAPI\YawGLAPI.csproj" />
   </ItemGroup>
+  <Import Project="..\SharedLib\SharedLib.projitems" Label="Shared" />
 </Project>
